Extract Action generator to its own subcommands undername space Action
