+++ lib/lotus/commands/action/new/action.rb.tt
