+++ lib/lotus/commands/action/new/action_without_view.rb.tt
