+++ lib/lotus/commands/action/new/view.rb.tt
