+++ lib/lotus/commands/action/new/view_spec.minitest.tt
