+++ lib/lotus/commands/action/new/view_spec.rspec.tt
